✨ feat(symfony): added symfony generator + refacto some files for better structure
