+++ Cargo.toml
@@ -1,13 +1,14 @@
 [package]
 name = "utilsgen"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
-repository = "https://github.com/anthodev/utilsgen"
+repository = "https://github.com/anthodev/utilsgen-rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ulid = "1.1.2"
-uuid = { version = "1.7.0", features = ["v4"] }
-chrono = "0.4.34"
-regex = "1.10.3"
+ulid = "1.1.*"
+uuid = { version = "1.7.*", features = ["v4"] }
+chrono = "0.4.*"
+regex = "1.10.*"
+rand = "0.8.*"
